LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-get-set-range-expected.txt


FAIL Getting and setting the animation range assert_equals: Initial value for rangeStart expected (string) "normal" but got (object) object "[object Object]"
